Write out metadata as UTF-8-encoded text
When writing a text file, if an encoding isn't specified, the platform
default will be used. This caused errors when gear/photo/activity
metadata contained characters that couldn't be encoded using that encoding.

For example, emojis cannot be written using the Windows default encoding
of `cp1253` so attempting to write out metadata that contained one would
raise exceptions.

This commit fixes this issue by explicitly asking Python to use UTF-8 to
encode all the text.