activity backup: also include total_elevation_gain, average_speed and max_speed
The unit of speed is typically m/s, so also accept a list of units in
valid_unit(): if all items of the list are valid, then consider the list
as valid, too.

With this, all the 6 stats included for an activity in the mobile app
are also tracked in .meta.json.