Improve method of updating the config file in-place
The `configparser` in the stdlib doesn't preserve comments when
reading/writing config files. Therefore, when the refresh token in the
config needed to be updated, the raw text of the config file was
modified using a regular expression.

This commits improves on this by using the `commented-configparser`
package. It will read all the comments from the file, then, when writing
the updated config, is able to preserve those comments.